requirements.txt
streamlit
google.generativeai
google.cloud.bigquery
pandas
db_dtypes
plotly
matplotlib